[Decomp][SpellMgr] Cohesion split: extract spell chain loaders to SpellMgrChains.cpp
