+++ src/game/Object/SpellMgr.cpp
@@ -3283,623 +3283,10 @@ SpellEntry const* SpellMgr::SelectAuraRankForLevel(SpellEntry const* spellInfo, 
     return NULL;
 }
 
-typedef UNORDERED_MAP<uint32, uint32> AbilitySpellPrevMap;
 
-/**
- * @brief Recursively builds spell chain links from ability rank data.
- *
- * @param chainMap The spell chain map being populated.
- * @param prevRanks The previous-rank lookup table.
- * @param spell_id The current spell id.
- * @param prev_id The previous spell id in the chain.
- * @param deep The remaining recursion depth guard.
- */
-static void LoadSpellChains_AbilityHelper(SpellChainMap& chainMap, AbilitySpellPrevMap const& prevRanks, uint32 spell_id, uint32 prev_id, uint32 deep = 30)
-{
-    // spell already listed in chains store
-    SpellChainMap::const_iterator chain_itr = chainMap.find(spell_id);
-    if (chain_itr != chainMap.end())
-    {
-        MANGOS_ASSERT(chain_itr->second.prev == prev_id);       // LoadSpellChains_AbilityHelper: Conflicting data in talents or spell abilities dbc
-        return;
-    }
 
-    // prev rank listed in main chain table (can fill correct data directly)
-    SpellChainMap::const_iterator prev_chain_itr = chainMap.find(prev_id);
-    if (prev_chain_itr != chainMap.end())
-    {
-        SpellChainNode node;
-        node.prev  = prev_id;
-        node.first = prev_chain_itr->second.first;
-        node.rank  = prev_chain_itr->second.rank + 1;
-        node.req   = 0;
-        chainMap[spell_id] = node;
-        return;
-    }
 
-    // prev spell not listed in prev ranks store, so it first rank
-    AbilitySpellPrevMap::const_iterator prev_itr = prevRanks.find(prev_id);
-    if (prev_itr == prevRanks.end())
-    {
-        SpellChainNode prev_node;
-        prev_node.prev  = 0;
-        prev_node.first = prev_id;
-        prev_node.rank  = 1;
-        prev_node.req   = 0;
-        chainMap[prev_id] = prev_node;
 
-        SpellChainNode node;
-        node.prev  = prev_id;
-        node.first = prev_id;
-        node.rank  = 2;
-        node.req   = 0;
-        chainMap[spell_id] = node;
-        return;
-    }
-
-    if (deep == 0)
-    {
-        MANGOS_ASSERT(false);   // LoadSpellChains_AbilityHelper: Infinity cycle in spell ability data
-        return;
-    }
-
-    // prev rank listed, so process it first
-    LoadSpellChains_AbilityHelper(chainMap, prevRanks, prev_id, prev_itr->second, deep - 1);
-
-    // prev rank must be listed now
-    prev_chain_itr = chainMap.find(prev_id);
-    if (prev_chain_itr == chainMap.end())
-    {
-        return;
-    }
-
-    SpellChainNode node;
-    node.prev  = prev_id;
-    node.first = prev_chain_itr->second.first;
-    node.rank  = prev_chain_itr->second.rank + 1;
-    node.req   = 0;
-    chainMap[spell_id] = node;
-}
-
-/**
- * @brief Loads spell rank chain data from DBC and database sources.
- */
-void SpellMgr::LoadSpellChains()
-{
-    mSpellChains.clear();                                   // need for reload case
-    mSpellChainsNext.clear();                               // need for reload case
-
-    // load known data for talents
-    for (unsigned int i = 0; i < sTalentStore.GetNumRows(); ++i)
-    {
-        TalentEntry const* talentInfo = sTalentStore.LookupEntry(i);
-        if (!talentInfo)
-        {
-            continue;
-        }
-
-        // not add ranks for 1 ranks talents (if exist non ranks spells then it will included in table data)
-        if (!talentInfo->RankID[1])
-        {
-            continue;
-        }
-
-        for (int j = 0; j < MAX_TALENT_RANK; ++j)
-        {
-            uint32 spell_id = talentInfo->RankID[j];
-            if (!spell_id)
-            {
-                continue;
-            }
-
-            if (!sSpellStore.LookupEntry(spell_id))
-            {
-                // sLog.outErrorDb("Talent %u not exist as spell",spell_id);
-                continue;
-            }
-
-            SpellChainNode node;
-            node.prev  = (j > 0) ? talentInfo->RankID[j - 1] : 0;
-            node.first = talentInfo->RankID[0];
-            node.rank  = j + 1;
-            node.req   = 0;
-
-            mSpellChains[spell_id] = node;
-        }
-    }
-
-    // load known data from spell abilities
-    {
-        // we can calculate ranks only after full data generation
-        AbilitySpellPrevMap prevRanks;
-        for (SkillLineAbilityMap::const_iterator ab_itr = mSkillLineAbilityMap.begin(); ab_itr != mSkillLineAbilityMap.end(); ++ab_itr)
-        {
-            uint32 spell_id = ab_itr->first;
-
-            // skip GM/test/internal spells.begin Its not have ranks anyway
-            if (ab_itr->second->skillId == SKILL_INTERNAL)
-            {
-                continue;
-            }
-
-            // some forward spells not exist and can be ignored (some outdated data)
-            SpellEntry const* spell_entry = sSpellStore.LookupEntry(spell_id);
-            if (!spell_entry)                               // no cases
-            {
-                continue;
-            }
-
-            // ignore spell without forwards (non ranked or missing info in skill abilities)
-            uint32 forward_id = ab_itr->second->forward_spellid;
-
-            // by some strange reason < 3.x clients not have forward spell for 2366
-            if (spell_id == 2366)                           // Herb Gathering, Apprentice
-            {
-                forward_id = 2368;
-            }
-
-            if (!forward_id)
-            {
-                continue;
-            }
-
-            // some forward spells not exist and can be ignored (some outdated data)
-            SpellEntry const* forward_entry = sSpellStore.LookupEntry(forward_id);
-            if (!forward_entry)
-            {
-                continue;
-            }
-
-            // some forward spells still exist but excluded from real use as ranks and not listed in skill abilities now
-            SkillLineAbilityMapBounds bounds = mSkillLineAbilityMap.equal_range(forward_id);
-            if (bounds.first == bounds.second)
-            {
-                continue;
-            }
-
-            // spell already listed in chains store
-            SpellChainMap::const_iterator chain_itr = mSpellChains.find(forward_id);
-            if (chain_itr != mSpellChains.end())
-            {
-                MANGOS_ASSERT(chain_itr->second.prev == spell_id);      // Conflicting data in talents or spell abilities dbc
-                continue;
-            }
-
-            // spell already listed in prev ranks store
-            AbilitySpellPrevMap::const_iterator prev_itr = prevRanks.find(forward_id);
-            if (prev_itr != prevRanks.end())
-            {
-                MANGOS_ASSERT(prev_itr->second == spell_id);            // Conflicting data in talents or spell abilities dbc
-                continue;
-            }
-
-            // prev rank listed in main chain table (can fill correct data directly)
-            SpellChainMap::const_iterator prev_chain_itr = mSpellChains.find(spell_id);
-            if (prev_chain_itr != mSpellChains.end())
-            {
-                SpellChainNode node;
-                node.prev  = spell_id;
-                node.first = prev_chain_itr->second.first;
-                node.rank  = prev_chain_itr->second.rank + 1;
-                node.req   = 0;
-
-                mSpellChains[forward_id] = node;
-                continue;
-            }
-
-            // need temporary store for later rank calculation
-            prevRanks[forward_id] = spell_id;
-        }
-
-        while (!prevRanks.empty())
-        {
-            uint32 spell_id = prevRanks.begin()->first;
-            uint32 prev_id  = prevRanks.begin()->second;
-            prevRanks.erase(prevRanks.begin());
-
-            LoadSpellChains_AbilityHelper(mSpellChains, prevRanks, spell_id, prev_id);
-        }
-    }
-
-    // load custom case
-    QueryResult* result = WorldDatabase.Query("SELECT `spell_id`, `prev_spell`, `first_spell`, `rank`, `req_spell` FROM `spell_chain`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-
-        sLog.outString(">> Loaded 0 spell chain records");
-        sLog.outErrorDb("`spell_chains` table is empty!");
-        sLog.outString();
-        return;
-    }
-
-    uint32 dbc_count = mSpellChains.size();
-    uint32 new_count = 0;
-    uint32 req_count = 0;
-
-    BarGoLink bar(result->GetRowCount());
-    do
-    {
-        bar.step();
-        Field* fields = result->Fetch();
-
-        uint32 spell_id = fields[0].GetUInt32();
-
-        SpellChainNode node;
-        node.prev  = fields[1].GetUInt32();
-        node.first = fields[2].GetUInt32();
-        node.rank  = fields[3].GetUInt8();
-        node.req   = fields[4].GetUInt32();
-
-        if (!sSpellStore.LookupEntry(spell_id))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_chain` does not exist", spell_id);
-            continue;
-        }
-
-        SpellChainMap::iterator chain_itr = mSpellChains.find(spell_id);
-        if (chain_itr != mSpellChains.end())
-        {
-            if (chain_itr->second.rank != node.rank)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` expected rank %u by DBC data.",
-                                spell_id, node.prev, node.first, node.rank, node.req, chain_itr->second.rank);
-                continue;
-            }
-
-            if (chain_itr->second.prev != node.prev)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` expected prev %u by DBC data.",
-                                spell_id, node.prev, node.first, node.rank, node.req, chain_itr->second.prev);
-                continue;
-            }
-
-            if (chain_itr->second.first != node.first)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` expected first %u by DBC data.",
-                                spell_id, node.prev, node.first, node.rank, node.req, chain_itr->second.first);
-                continue;
-            }
-
-            // update req field by table data
-            if (node.req)
-            {
-                chain_itr->second.req = node.req;
-                ++req_count;
-                continue;
-            }
-
-            // in other case redundant
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) already added (talent or spell ability with forward) and non need in `spell_chain`",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        if (node.prev != 0 && !sSpellStore.LookupEntry(node.prev))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has nonexistent previous rank spell.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        if (!sSpellStore.LookupEntry(node.first))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not existing first rank spell.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        // check basic spell chain data integrity (note: rank can be equal 0 or 1 for first/single spell)
-        if ((spell_id == node.first) != (node.rank <= 1) ||
-            (spell_id == node.first) != (node.prev == 0) ||
-            (node.rank <= 1) != (node.prev == 0))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not compatible chain data.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        if (node.req != 0 && !sSpellStore.LookupEntry(node.req))
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not existing required spell.",
-                            spell_id, node.prev, node.first, node.rank, node.req);
-            continue;
-        }
-
-        // talents not required data in spell chain for work, but must be checked if present for integrity
-        if (TalentSpellPos const* pos = GetTalentSpellPos(spell_id))
-        {
-            if (node.rank != pos->rank + 1)
-            {
-                sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong rank.",
-                                spell_id, node.prev, node.first, node.rank, node.req);
-                continue;
-            }
-
-            if (TalentEntry const* talentEntry = sTalentStore.LookupEntry(pos->talent_id))
-            {
-                if (node.first != talentEntry->RankID[0])
-                {
-                    sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong first rank spell.",
-                                    spell_id, node.prev, node.first, node.rank, node.req);
-                    continue;
-                }
-
-                if (node.rank > 1 && node.prev != talentEntry->RankID[node.rank - 1 - 1])
-                {
-                    sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong prev rank spell.",
-                                    spell_id, node.prev, node.first, node.rank, node.req);
-                    continue;
-                }
-
-                if (node.req != talentEntry->DependsOnSpell)
-                {
-                    sLog.outErrorDb("Talent %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has wrong required spell.",
-                                    spell_id, node.prev, node.first, node.rank, node.req);
-                    continue;
-                }
-            }
-        }
-
-        mSpellChains[spell_id] = node;
-
-        ++new_count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    // additional integrity checks
-    for (SpellChainMap::const_iterator i = mSpellChains.begin(); i != mSpellChains.end(); ++i)
-    {
-        if (i->second.prev)
-        {
-            SpellChainMap::const_iterator i_prev = mSpellChains.find(i->second.prev);
-            if (i_prev == mSpellChains.end())
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not found previous rank spell in table.",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req);
-            }
-            else if (i_prev->second.first != i->second.first)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has different first spell in chain compared to previous rank spell (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_prev->second.prev, i_prev->second.first, i_prev->second.rank, i_prev->second.req);
-            }
-            else if (i_prev->second.rank + 1 != i->second.rank)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has different rank compared to previous rank spell (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_prev->second.prev, i_prev->second.first, i_prev->second.rank, i_prev->second.req);
-            }
-        }
-
-        if (i->second.req)
-        {
-            SpellChainMap::const_iterator i_req = mSpellChains.find(i->second.req);
-            if (i_req == mSpellChains.end())
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has not found required rank spell in table.",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req);
-            }
-            else if (i_req->second.first == i->second.first)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has required rank spell from same spell chain (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_req->second.prev, i_req->second.first, i_req->second.rank, i_req->second.req);
-            }
-            else if (i_req->second.req)
-            {
-                sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has required rank spell with required spell (prev: %u, first: %u, rank: %d, req: %u).",
-                                i->first, i->second.prev, i->second.first, i->second.rank, i->second.req,
-                                i_req->second.prev, i_req->second.first, i_req->second.rank, i_req->second.req);
-            }
-        }
-    }
-
-    // fill next rank cache
-    for (SpellChainMap::const_iterator i = mSpellChains.begin(); i != mSpellChains.end(); ++i)
-    {
-        uint32 spell_id = i->first;
-        SpellChainNode const& node = i->second;
-
-        if (node.prev)
-        {
-            mSpellChainsNext.insert(SpellChainMapNext::value_type(node.prev, spell_id));
-        }
-
-        if (node.req)
-        {
-            mSpellChainsNext.insert(SpellChainMapNext::value_type(node.req, spell_id));
-        }
-    }
-
-    // check single rank redundant cases (single rank talents not added by default so this can be only custom cases)
-    for (SpellChainMap::const_iterator i = mSpellChains.begin(); i != mSpellChains.end(); ++i)
-    {
-        // skip non-first ranks, and spells with additional reqs
-        if (i->second.rank > 1 || i->second.req)
-        {
-            continue;
-        }
-
-        if (mSpellChainsNext.find(i->first) == mSpellChainsNext.end())
-        {
-            sLog.outErrorDb("Spell %u (prev: %u, first: %u, rank: %d, req: %u) listed in `spell_chain` has single rank data, so redundant.",
-                            i->first, i->second.prev, i->second.first, i->second.rank, i->second.req);
-        }
-    }
-
-    sLog.outString(">> Loaded %u spell chain records (%u from DBC data with %u req field updates, and %u loaded from table)", dbc_count + new_count, dbc_count, req_count, new_count);
-    sLog.outString();
-}
-
-/**
- * @brief Builds the spell-to-skill-learning map from DBC data.
- */
-void SpellMgr::LoadSpellLearnSkills()
-{
-    mSpellLearnSkills.clear();                              // need for reload case
-
-    // search auto-learned skills and add its to map also for use in unlearn spells/talents
-    uint32 dbc_count = 0;
-    BarGoLink bar(sSpellStore.GetNumRows());
-    for (uint32 spell = 0; spell < sSpellStore.GetNumRows(); ++spell)
-    {
-        bar.step();
-        SpellEntry const* entry = sSpellStore.LookupEntry(spell);
-
-        if (!entry)
-        {
-            continue;
-        }
-
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            if (entry->Effect[i] == SPELL_EFFECT_SKILL)
-            {
-                SpellLearnSkillNode dbc_node;
-                dbc_node.skill    = entry->EffectMiscValue[i];
-                dbc_node.step     = entry->CalculateSimpleValue(SpellEffectIndex(i));
-                if (dbc_node.skill != SKILL_RIDING)
-                {
-                    dbc_node.value = 1;
-                }
-                else
-                {
-                    dbc_node.value = dbc_node.step * 75;
-                }
-                dbc_node.maxvalue = dbc_node.step * 75;
-
-                mSpellLearnSkills[spell] = dbc_node;
-                ++dbc_count;
-                break;
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded %u Spell Learn Skills from DBC", dbc_count);
-    sLog.outString();
-}
-
-/**
- * @brief Loads explicit and DBC-derived learned-spell relationships.
- */
-void SpellMgr::LoadSpellLearnSpells()
-{
-    mSpellLearnSpells.clear();                              // need for reload case
-
-    //                                                0      1        2
-    QueryResult* result = WorldDatabase.Query("SELECT `entry`, `SpellID`, `Active` FROM `spell_learn_spell`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-
-        sLog.outString();
-        sLog.outString(">> Loaded 0 spell learn spells");
-        sLog.outErrorDb("`spell_learn_spell` table is empty!");
-        return;
-    }
-
-    uint32 count = 0;
-
-    BarGoLink bar(result->GetRowCount());
-    do
-    {
-        bar.step();
-        Field* fields = result->Fetch();
-
-        uint32 spell_id    = fields[0].GetUInt32();
-
-        SpellLearnSpellNode node;
-        node.spell      = fields[1].GetUInt32();
-        node.active     = fields[2].GetBool();
-        node.autoLearned = false;
-
-        if (!sSpellStore.LookupEntry(spell_id))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_learn_spell` does not exist", spell_id);
-            continue;
-        }
-
-        if (!sSpellStore.LookupEntry(node.spell))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_learn_spell` learning nonexistent spell %u", spell_id, node.spell);
-            continue;
-        }
-
-        if (GetTalentSpellCost(node.spell))
-        {
-            sLog.outErrorDb("Spell %u listed in `spell_learn_spell` attempt learning talent spell %u, skipped", spell_id, node.spell);
-            continue;
-        }
-
-        mSpellLearnSpells.insert(SpellLearnSpellMap::value_type(spell_id, node));
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    // search auto-learned spells and add its to map also for use in unlearn spells/talents
-    uint32 dbc_count = 0;
-    for (uint32 spell = 0; spell < sSpellStore.GetNumRows(); ++spell)
-    {
-        SpellEntry const* entry = sSpellStore.LookupEntry(spell);
-
-        if (!entry)
-        {
-            continue;
-        }
-
-        for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-        {
-            if (entry->Effect[i] == SPELL_EFFECT_LEARN_SPELL)
-            {
-                SpellLearnSpellNode dbc_node;
-                dbc_node.spell       = entry->EffectTriggerSpell[i];
-                dbc_node.active      = true;                // all dbc based learned spells is active (show in spell book or hide by client itself)
-
-                // ignore learning nonexistent spells (broken/outdated/or generic learning spell 483
-                if (!sSpellStore.LookupEntry(dbc_node.spell))
-                {
-                    continue;
-                }
-
-                // talent or passive spells or skill-step spells auto-casted and not need dependent learning,
-                // pet teaching spells don't must be dependent learning (casted)
-                // other required explicit dependent learning
-                dbc_node.autoLearned = entry->EffectImplicitTargetA[i] == TARGET_PET || GetTalentSpellCost(spell) > 0 || IsPassiveSpell(entry) || IsSpellHaveEffect(entry, SPELL_EFFECT_SKILL_STEP);
-
-                SpellLearnSpellMapBounds db_node_bounds = GetSpellLearnSpellMapBounds(spell);
-
-                bool found = false;
-                for (SpellLearnSpellMap::const_iterator itr = db_node_bounds.first; itr != db_node_bounds.second; ++itr)
-                {
-                    if (itr->second.spell == dbc_node.spell)
-                    {
-                        sLog.outErrorDb("Spell %u auto-learn spell %u in spell.dbc then the record in `spell_learn_spell` is redundant, please fix DB.",
-                                        spell, dbc_node.spell);
-                        found = true;
-                        break;
-                    }
-                }
-
-                if (!found)                                 // add new spell-spell pair if not found
-                {
-                    mSpellLearnSpells.insert(SpellLearnSpellMap::value_type(spell, dbc_node));
-                    ++dbc_count;
-                }
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded %u spell learn spells + %u found in DBC", count, dbc_count);
-    sLog.outString();
-}
 
 /**
  * @brief Loads and validates scripted spell targets.
